+++ test/hash/minhash.js
@@ -1,0 +1,16 @@
+/**
+ * Talisman helpers/minhash tests
+ * ===============================
+ *
+ */
+// import assert from 'assert';
+// import createMinHash from '../../src/hash/minhash';
+// import seedrandom from 'seedrandom';
+
+// describe('minhash', function() {
+
+//   it('should produce the correct signature.', function() {
+//     const rng = seedrandom('shawarma'),
+//           minhash = createMinHash({rng, hashes: 128});
+//   });
+// });